+++ base_local_planner/CMakeLists.txt
@@ -39,6 +39,8 @@ add_message_files(
     Position2DInt.msg
 )
 
+catkin_python_setup()
+
 generate_messages(
     DEPENDENCIES
         std_msgs
